Assign weakself to self.

+++ FirebaseRemoteConfig/Sources/RCNConfigDBManager.m
@@ -761,11 +761,10 @@ static NSArray *RemoteConfigMetadataTableColumnsInOrder() {
 
     if (handler) {
       dispatch_async(dispatch_get_main_queue(), ^{
-        handler(
-            YES, @{
-              @RCNExperimentTableKeyPayload : [experimentPayloads copy],
-              @RCNExperimentTableKeyMetadata : [experimentMetadata copy]
-            });
+        handler(YES, @{
+          @RCNExperimentTableKeyPayload : [experimentPayloads copy],
+          @RCNExperimentTableKeyMetadata : [experimentMetadata copy]
+        });
       });
     }
   });
@@ -899,8 +898,12 @@ static NSArray *RemoteConfigMetadataTableColumnsInOrder() {
 - (void)deleteRecordFromMainTableWithNamespace:(NSString *)namespace_p
                               bundleIdentifier:(NSString *)bundleIdentifier
                                     fromSource:(RCNDBSource)source {
-  __weak RCNConfigDBManager *weakSelf;
+  __weak RCNConfigDBManager *weakSelf = self;
   dispatch_async(_databaseOperationQueue, ^{
+    RCNConfigDBManager *strongSelf = weakSelf;
+    if (!strongSelf) {
+      return;
+    }
     NSArray *params = @[ bundleIdentifier, namespace_p ];
     const char *SQL =
         "DELETE FROM " RCNTableNameMain " WHERE bundle_identifier = ? and namespace = ?";
@@ -909,42 +912,54 @@ static NSArray *RemoteConfigMetadataTableColumnsInOrder() {
     } else if (source == RCNDBSourceActive) {
       SQL = "DELETE FROM " RCNTableNameMainActive " WHERE bundle_identifier = ? and namespace = ?";
     }
-    [weakSelf executeQuery:SQL withParams:params];
+    [strongSelf executeQuery:SQL withParams:params];
   });
 }
 
 - (void)deleteRecordWithBundleIdentifier:(NSString *)bundleIdentifier
                             isInternalDB:(BOOL)isInternalDB {
-  __weak RCNConfigDBManager *weakSelf;
+  __weak RCNConfigDBManager *weakSelf = self;
   dispatch_async(_databaseOperationQueue, ^{
+    RCNConfigDBManager *strongSelf = weakSelf;
+    if (!strongSelf) {
+      return;
+    }
     const char *SQL = "DELETE FROM " RCNTableNameInternalMetadata " WHERE key LIKE ?";
     if (!isInternalDB) {
       SQL = "DELETE FROM " RCNTableNameMetadata " WHERE bundle_identifier = ?";
     }
     NSArray *params = @[ bundleIdentifier ];
-    [weakSelf executeQuery:SQL withParams:params];
+    [strongSelf executeQuery:SQL withParams:params];
   });
 }
 
 - (void)deleteAllRecordsFromTableWithSource:(RCNDBSource)source {
-  __weak RCNConfigDBManager *weakSelf;
+  __weak RCNConfigDBManager *weakSelf = self;
   dispatch_async(_databaseOperationQueue, ^{
+    RCNConfigDBManager *strongSelf = weakSelf;
+    if (!strongSelf) {
+      return;
+    }
     const char *SQL = "DELETE FROM " RCNTableNameMain;
     if (source == RCNDBSourceDefault) {
       SQL = "DELETE FROM " RCNTableNameMainDefault;
     } else if (source == RCNDBSourceActive) {
       SQL = "DELETE FROM " RCNTableNameMainActive;
     }
-    [weakSelf executeQuery:SQL];
+    [strongSelf executeQuery:SQL];
   });
 }
 
 - (void)deleteExperimentTableForKey:(NSString *)key {
-  __weak RCNConfigDBManager *weakSelf;
+  __weak RCNConfigDBManager *weakSelf = self;
+  RCNConfigDBManager *strongSelf = weakSelf;
+  if (!strongSelf) {
+    return;
+  }
   dispatch_async(_databaseOperationQueue, ^{
     NSArray *params = @[ key ];
     const char *SQL = "DELETE FROM " RCNTableNameExperiment " WHERE key = ?";
-    [weakSelf executeQuery:SQL withParams:params];
+    [strongSelf executeQuery:SQL withParams:params];
   });
 }
